Made the voting feature entities

+++ Blueboard/Features/Auth/Commands/DeleteUserGroup.cs
@@ -4,6 +4,7 @@ using Blueboard.Infrastructure.Persistence;
 using Blueboard.Infrastructure.Persistence.Entities;
 using Helpers.WebApi.Exceptions;
 using MediatR;
+using Microsoft.EntityFrameworkCore;
 
 namespace Blueboard.Features.Auth.Commands;
 
@@ -27,13 +28,17 @@ public static class DeleteUserGroup
 
         public async Task<Unit> Handle(Command request, CancellationToken cancellationToken)
         {
-            var userGroup = await _context.UserGroups.FindAsync(request.Id);
+            var userGroup = await _context.UserGroups.Include(g => g.ImageVotings)
+                .FirstOrDefaultAsync(g => g.Id == request.Id, cancellationToken);
 
             if (userGroup == null)
                 throw new NotFoundException(nameof(UserGroup), request.Id);
 
             if (userGroup.Id == AuthConstants.DefaultUserGroupID)
                 throw new BadRequestException("Nem törölheted az alapértelmezett csoportot.");
+
+            if (userGroup.ImageVotings.Count > 0)
+                throw new BadRequestException("Nem törölheted a csoportot, mert szavazás van hozzárendelve.");
 
             _context.UserGroups.Remove(userGroup);
             await _context.SaveChangesAsync(cancellationToken);
